Added an ID to routes registered to an HTTP endpoint
Fixes #2

Added an ID to the base route object. This ID will be generated when a
base route is created and kept for rest of the session. Routes are
generated in the order which routes are generated.
Also modified the `http::Request` to provide a method that can be used
to query the route ID.
When registering a route handler, the route ID will be returned.